Move Swig Generated classes to Wrapper Package
