+++ TransbankPosSDK/Wrapper/TransbankWrapPINVOKE.cs
@@ -8,7 +8,7 @@
 // the SWIG interface file instead.
 //------------------------------------------------------------------------------
 
-namespace Transbank.POS {
+namespace Transbank.POS.Wrapper {
 
 class TransbankWrapPINVOKE {
 
@@ -188,43 +188,37 @@ class TransbankWrapPINVOKE {
   }
 
 
-  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOS_TBK_OK_get___")]
-  public static extern int TBK_OK_get();
-
-  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOS_TBK_NOK_get___")]
-  public static extern int TBK_NOK_get();
-
-  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOS_print_ports___")]
+  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOSfWrapper_print_ports___")]
   public static extern void print_ports();
 
-  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOS_list_ports___")]
+  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOSfWrapper_list_ports___")]
   public static extern string list_ports();
 
-  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOS_get_configured_port_name___")]
+  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOSfWrapper_get_configured_port_name___")]
   public static extern string get_configured_port_name();
 
-  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOS_open_configured_port___")]
+  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOSfWrapper_open_configured_port___")]
   public static extern int open_configured_port();
 
-  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOS_select_port___")]
+  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOSfWrapper_select_port___")]
   public static extern int select_port(string jarg1);
 
-  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOS_configure_port___")]
-  public static extern int configure_port();
+  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOSfWrapper_configure_port___")]
+  public static extern int configure_port(int jarg1);
 
-  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOS_get_totals___")]
+  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOSfWrapper_get_totals___")]
   public static extern int get_totals();
 
-  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOS_load_keys___")]
+  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOSfWrapper_load_keys___")]
   public static extern int load_keys();
 
-  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOS_polling___")]
+  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOSfWrapper_polling___")]
   public static extern int polling();
 
-  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOS_set_normal_mode___")]
+  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOSfWrapper_set_normal_mode___")]
   public static extern int set_normal_mode();
 
-  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOS_close_port___")]
+  [global::System.Runtime.InteropServices.DllImport("TransbankWrap", EntryPoint="CSharp_TransbankfPOSfWrapper_close_port___")]
   public static extern int close_port();
 }
